requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
charset-normalizer==3.1.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
psycopg==3.1.8
psycopg-binary==3.1.8
PyYAML==6.0
typing_extensions==4.5.0
yarl==1.8.2
